server/requirements.txt
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.5.4
boto3==1.16.1
botocore==1.19.1
CacheControl==0.12.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.2
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-simplejwt==5.0.0
docutils==0.15.2
drf-social-oauth2==1.0.8
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==8.4.0
progress==1.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
